Merged PR 85: Added logic trap errors if target workitem are missing
Not 100% sure about this one, but it does provide resilience against poor/strange practices e.g. closing a test case that is in an active test plan.

It does not change the function for normal operations, but put in skip/continue with logging for edge cases

Added logic trap errors if target workitem are missing
The logic for migrating work items means that closes/done work items do not migrate. This cause a problem if they are refered to in test plans or as link. The changes trap for these conditions

Related work items: #4981