Status Dates Fixes (#275)
Allows the Closed Date, Activated Date and Resolved Date  to be changed.

I discovered that the value cannot be written in the revision where the change happens but in subsequent revisions this is not a problem.
To bypass the Closed Date issue a new revision needs to be added in the item to migrate. I was not able to do that in the code yet, but adding in the work item on the server to migrate ensures all date values are correctly passed.