feat: select build/release definitions to migrate
allow end user to select the build and release pipelines they want to migrate.
the configuration uses the existing string array properties on the configuration options.
added a flag to not query details for git repos when migrating pipelines. the details were not being used and would also cause the process to fail when it queried a disabled repo.
documentation for new method added and some updated.