Migrating a YAML pipeline should not throw exceptions (#1099)
* fix(): make repository clean nullable

exception would be thrown when migrating yaml pipelines because the `repositroy.clean` prop is always null

* fix(): add default collection values

yaml pipelines don't have the concept phases so those should be default to an empty collection instead of null.

* fix(): added conditional serialization

based on the type of pipeline (classic vs yaml) the process will have different values:
classic -> phases
yaml -> filename

this will make sure it serializes only the process needed when migrating to the target.