Update AzureDevOpsPipelineProcessor.cs
Added a check to see if there are any properties in a pipeline's repository and if there are no properties in the repository default values will be set. This is because if there are no properties set a NullReferenceException error occurs on line 242 "var sourceConnectedServiceId = definitionToBeMigrated.Repository.Properties.ConnectedServiceId;" due to there not being an instance of Properties as there is no property data in the pipeline's repository. The check fixes this issue, and a warning is also added to the log explaining to the user that property data is missing from the repository and default values have been set.