Update DetectVersionService.cs (#1684)
* Update DetectVersionService.cs

The package search will search the winget repository and will always return the newest version. Package.Version and Package.AvailableVersion will always be the same here.
I changed it to `GetInstalledPackages(packageId)` instead of `SearchPackage(PackageId)`. This will give you the correct installed and available version and should also be faster, because you are just going through the installed packages.

* Update DetectVersionService.cs

---------

Co-authored-by: Martin Hinshelwood nkdAgility.com <martin@nkdagility.com>