Update version in log (#1694)
* Figuring it out

* Update to both instal and update the package... tests

* Added the ability to both install and update the appliation using Winget. This should run on Windows Server by skipping all the checks as winget is not supported ,we still deploy to chocolaty.

* Update DetectVersionService.cs (#1684)

* Update DetectVersionService.cs

The package search will search the winget repository and will always return the newest version. Package.Version and Package.AvailableVersion will always be the same here.
I changed it to `GetInstalledPackages(packageId)` instead of `SearchPackage(PackageId)`. This will give you the correct installed and available version and should also be faster, because you are just going through the installed packages.

* Update DetectVersionService.cs

---------

Co-authored-by: Martin Hinshelwood nkdAgility.com <martin@nkdagility.com>

* Added version number to log so we can see it in partial logs as well

* Update to take it out of the mane code

---------

Co-authored-by: Bjarne <59922570+basicx-StrgV@users.noreply.github.com>