Update all dependanncies 2024-02-18 (#1859)
There was an error in https://github.com/basicx-StrgV/WGet.NET/issues/32 so we are taking the opportunity to update all of the dependencies.

This update also includes a refactor of the `DetectVersionService2` but I don't believe that it has resolved the issue. We have methods being fired from threads that I cant track down.