[preview] Update to documenation and code to make  (#1960)
It looks like the follwoing was not clear and continues to result in
support requests:

> There are 4 Nodes (Area or Iteration) found in the history of the
Source that are missing from the Target! These MUST be added or mapped
before we can continue using the instructions on
>
https://nkdagility.com/learn/azure-devops-migration-tools/Reference/v1/Processors/WorkItemMigrationContext/#iteration-maps-and-area-maps

Trying with the following:

> !! There are MISSING Area or Iteration Paths"
> NOTE: It is NOT possible to migrate a work item if the Area or
Iteration path does not exist on the target project. This is because the
work item will be created with the same Area and Iteration path as the
source work item with the project name swapped. The work item will not
be created if the path does not exist. The only way to resolve this is
to follow the instructions:
> !! There are {missingAreaPaths} Nodes (Area or Iteration) found in the
history of the Source that need to be added to the Target! These MUST be
added or mapped before we can continue using the instructions on
https://nkdagility.com/learn/azure-devops-migration-tools//Reference/v2/ProcessorEnrichers/TfsNodeStructure/#iteration-maps-and-area-maps

Lets see if this reduces these requests...