📝 (docs): update MigrationTools and VstsSyncMigrator.Core documentation
Add new members to `MigrationTools.xml` for `WorkItemBulkEditProcessorConfig` properties. Correct the class and options class file paths in `workitembulkeditprocessor.yaml` and `workitembulkeditprocessor.md`. Swap the descriptions for `WorkItemDelete` and `WorkItemBulkEditProcessor` in `VstsSyncMigrator.Core.xml`.

These changes ensure that the documentation accurately reflects the current state of the codebase, providing correct paths and descriptions for better clarity and usability.